plugin: added "Show in filelist" option to context menu
When you go to context menu when you're on entry in playlist
you can select "show in filelist" option which will
show contents of directory of this entry in filelist.

This can come handy for example to quickly find out where
is file located or to add other related entries in directory
to playlist.